+++ colossalai/zero/__init__.py
@@ -1,1 +1,8 @@
+from .gemini import GeminiAdamOptimizer, GeminiDDP, ZeroDDP, ZeroOptimizer
 from .low_level import LowLevelZeroOptimizer
+from .wrapper import zero_model_wrapper, zero_optim_wrapper
+
+__all__ = [
+    'ZeroDDP', 'GeminiDDP', 'ZeroOptimizer', 'GeminiAdamOptimizer', 'zero_model_wrapper', 'zero_optim_wrapper',
+    'LowLevelZeroOptimizer'
+]
