+++ colossalai/zero/gemini/__init__.py
@@ -1,4 +1,10 @@
 from .chunk import ChunkManager, TensorInfo, TensorState, search_chunk_configuration
+from .gemini_ddp import GeminiDDP, ZeroDDP
 from .gemini_mgr import GeminiManager
+from .gemini_optimizer import GeminiAdamOptimizer, ZeroOptimizer
+from .utils import get_static_torch_model
 
-__all__ = ['GeminiManager', 'TensorInfo', 'TensorState', 'ChunkManager', 'search_chunk_configuration']
+__all__ = [
+    'GeminiManager', 'TensorInfo', 'TensorState', 'ChunkManager', 'search_chunk_configuration', 'ZeroDDP', 'GeminiDDP',
+    'get_static_torch_model', 'GeminiAdamOptimizer', 'ZeroOptimizer'
+]
