+++ tests/test_gemini/update/test_get_torch_model.py
@@ -6,13 +6,13 @@ import torch
 import torch.multiprocessing as mp
 
 import colossalai
-from colossalai.nn.parallel import GeminiDDP
-from colossalai.nn.parallel.utils import get_static_torch_model
 from colossalai.tensor import ColoParameter
 from colossalai.testing import parameterize, rerun_if_address_is_in_use
 from colossalai.utils import free_port
 from colossalai.utils.cuda import get_current_device
 from colossalai.utils.model.colo_init_context import ColoInitContext
+from colossalai.zero.gemini import GeminiDDP
+from colossalai.zero.gemini.utils import get_static_torch_model
 from tests.components_to_test.registry import non_distributed_component_funcs
